refactor: replace LoggingService with getLogService for improved logging functionality
